Refactoring to use logging framework
To implement OutputEvents we may need to capture the stdout and stderr
of the debuggee. However, if the DAP server is also outputting to stdout
and stderr its messages will be mixed up with the debuggee's.

This commit introduces a logging action to ensure it is possible to
redirect all of the DAP server's output to a particular handle, thereby
separating it from the debuggee.

Fixes #9